Use oklab color conversion functions.
Refactor color conversions into one ColorUtils.dll.
