+++ misc/package/Data Files/MWSE/core/mwse/ui/tes3uiElement/createColorPicker/colorUtils/src/types.h
@@ -1,0 +1,13 @@
+#pragma once
+
+struct RGB {
+	float r;
+	float g;
+	float b;
+};
+
+struct HSV {
+	float h;
+	float s;
+	float v;
+};
