--- misc/package/Data Files/MWSE/core/mwse/ui/tes3uiElement/createColorPicker/oklab/init.lua
@@ -1,137 +1,0 @@
-local ffi = require("ffi")
-
-ffi.cdef[[
-	typedef struct {
-		float r;
-		float g;
-		float b;
-	} RGB;
-
-	typedef struct {
-		float h;
-		float s;
-		float v;
-	} HSV;
-	RGB okhsv_to_srgb(HSV hsv);
-	HSV srgb_to_okhsv(RGB rgb);
-	RGB hsv_to_srgb(HSV c);
-	HSV srgb_to_hsv(RGB rgb);
-	void generate_main_picker(const double hue, RGB *data, const uint32_t width, const uint32_t height);
-	void generate_preview_image(const RGB color, const double alpha, RGB *destination, double *alphas, RGB *checkers, const int32_t width, const uint32_t height);
-	void generate_saturation_picker(const double hue, const double value, RGB *data, const uint32_t width, const uint32_t height);
-]]
-local oklab = ffi.load(".\\Data Files\\MWSE\\core\\mwse\\ui\\tes3uiElement\\createColorPicker\\oklab\\liboklab")
-local hsvlib = ffi.load(".\\Data Files\\MWSE\\core\\mwse\\ui\\tes3uiElement\\createColorPicker\\oklab\\libhsv")
-
---- @class ffiImagePixel : ffi.cdata*
---- @field r number Red in range [0, 1].
---- @field g number Green in range [0, 1].
---- @field b number Blue in range [0, 1].
-
---- @class ffiHSV : ffi.cdata*
---- @field h number Hue in range [0, 360)
---- @field s number Saturation in range [0, 1]
---- @field v number Value/brightness in range [0, 1]
-
-
---- @class liboklab
---- @field hsvtosrgb fun(hsv: HSV): ImagePixel
---- @field srgbtohsv fun(rgb: ImagePixelArgument|PremulImagePixelA): HSV
---- @field hsvlib_hsv_to_srgb fun(c: ffiHSV): ffiImagePixel
---- @field hsvlib_srgb_to_hsv fun(rgb: ffiImagePixel): ffiHSV
-local this = {}
-
-this.hsvlib_hsv_to_srgb = hsvlib.hsv_to_srgb
-this.hsvlib_srgb_to_hsv = hsvlib.srgb_to_hsv
-
---- @param hue number
---- @param image Image
-function this.generate_image(hue, image)
-	hsvlib.generate_main_picker(hue, image.data, image.width, image.height)
-end
-
---- @param hue number Hue in range [0, 360)
---- @param value number Value in range [0, 1]
---- @param image Image
-function this.generate_saturation_bar(hue, value, image)
-	hsvlib.generate_saturation_picker(hue, value, image.data, image.width, image.height);
-end
-
---- @param color ffiImagePixel
---- @param alpha number
---- @param destination Image
---- @param checkers Image
-function this.generate_preview(color, alpha, destination, checkers)
-	hsvlib.generate_preview_image(
-		color, alpha,
-		destination.data, destination.alphas, checkers.data,
-		destination.width, destination.height
-	)
-end
-
-function this.hsvtosrgb(hsv)
-	local arg = ffi.new("HSV")
-	--- @diagnostic disable: inject-field
-	arg.h = hsv.h
-	arg.s = hsv.s
-	arg.v = hsv.v
-	--- @diagnostic enable: inject-field
-	local ret = hsvlib.hsv_to_srgb(arg)
-	return {
-		r = ret.r,
-		g = ret.g,
-		b = ret.b,
-	}
-end
-
---- @param rgb ImagePixelArgument|PremulImagePixelA
-function this.srgbtohsv(rgb)
-	local arg = ffi.new("RGB")
-	--- @diagnostic disable: inject-field
-	arg.r = rgb.r
-	arg.g = rgb.g
-	arg.b = rgb.b
-	--- @diagnostic enable: inject-field
-	local ret = hsvlib.srgb_to_hsv(arg)
-	return {
-		h = ret.h,
-		s = ret.s,
-		v = ret.v,
-	}
-end
-
---- @param hsv HSV
-function this.oklab_hsvtosrgb(hsv)
-	local arg = ffi.new("HSV")
-	--- @diagnostic disable: inject-field
-	-- okhsv_to_srgb h is in [0, 1] range.
-	arg.h = hsv.h / 360
-	arg.s = hsv.s
-	arg.v = hsv.v
-	--- @diagnostic enable: inject-field
-	local ret = oklab.okhsv_to_srgb(arg)
-	return {
-		r = ret.r,
-		g = ret.g,
-		b = ret.b,
-	}
-end
-
---- @param rgb ImagePixelArgument|PremulImagePixelA
-function this.oklab_srgbtohsv(rgb)
-	local arg = ffi.new("RGB")
-	--- @diagnostic disable: inject-field
-	arg.r = rgb.r
-	arg.g = rgb.g
-	arg.b = rgb.b
-	--- @diagnostic enable: inject-field
-	local ret = oklab.srgb_to_okhsv(arg)
-	return {
-		-- srgb_to_okhsv returns h in [0, 1] range.
-		h = ret.h * 360,
-		s = ret.s,
-		v = ret.v,
-	}
-end
-
-return this
